Update AmityUIKit v3.0.0

+++ UpstraUIKit/UpstraUIKit/Modules/Comunity/Category List/AmityCategoryListScreenViewModel.swift
@@ -35,7 +35,7 @@ class AmityCategoryListScreenViewModel: AmityCategoryListScreenViewModelType {
     }
     
     func item(at indexPath: IndexPath) -> AmityCommunityCategory? {
-        return categoryCollection?.object(at: UInt(indexPath.row))
+        return categoryCollection?.object(at: indexPath.row)
     }
     
     func loadNext() {
